+++ src/main/exceptions/TweetGenerationException.ts
@@ -1,6 +1,0 @@
-export class TweetGenerationException extends Error {
-    constructor(message: string) {
-        super(message);
-        this.name = 'TweetGenerationException';
-    }
-}