+++ src/main/services/movie.service.ts
@@ -1,16 +1,16 @@
-import { Movie } from '../models/movie.model';
+import { MovieDTO } from '../models/movie.model';
 import { IMovieService } from '../interfaces/movie.service.interface';
 import { FileReader } from '../utils/fileReader';
 import logger from '../config/logger';
 import { FileReadException } from '../exceptions/FileReadException';
 
 export class MovieService implements IMovieService {
-    private movies: Movie[];
-    private movieMap: Map<string, Movie>;
+    private movies: MovieDTO[];
+    private movieMap: Map<string, MovieDTO>;
 
     constructor(moviesFile: string) {
         try {
-            this.movies = FileReader.readJSON<Movie[]>(moviesFile);
+            this.movies = FileReader.readJSON<MovieDTO[]>(moviesFile);
             this.movieMap = new Map(this.movies.map(movie => [movie.title, movie]));
         } catch (error) {
             logger.error(`Error reading movies file: ${error.message}`);
@@ -23,7 +23,7 @@ export class MovieService implements IMovieService {
      * @param movieId The title of the movie to find.
      * @returns The movie object if found, otherwise undefined.
      */
-    findMovieById(movieId: string): Movie | undefined {
+    findMovieById(movieId: string): MovieDTO | undefined {
         return this.movieMap.get(movieId);
     }
 }