+++ CHANGELOG.md
@@ -4,5 +4,9 @@
 ## [0.1.0]
 
 ### Added
+
+- Refactored the main module and added unit test cases
+- Added interfaces,constants and modularized the code
+- Added custom exceptions
 - Added service basic implementation
 - Initial configuration of the project with TypeScript setup.